+++ src/output.c
@@ -138,10 +138,12 @@ PrintUpdate (int rep, int *Geno, int *PreGeno, double *Alpha, double *Fst, doubl
   /*print a bunch of stuff to screen during run: rep, alpha, f, KLD, likelihood...
     Also occasionally print a header banner to define the variables. */
 
-  double logprob=0.0;
+  double logprob, like2;
   /*  int i;
    *  int printalign; */
   int pop;
+	
+	logprob=0.0;
 
   if ((rep < BURNIN + UPDATEFREQ) && (rep > BURNIN)) {
     printf ("\nBURNIN completed");
@@ -170,32 +172,21 @@ PrintUpdate (int rep, int *Geno, int *PreGeno, double *Alpha, double *Fst, doubl
     }
   }
 
-  
-	printf ("here -1");
-  
   /*calculate some stuff */
   if (LINKAGE) {
-			printf ("here 0");
-		
     if (rep <= ADMBURNIN) {
-			printf ("here 1");
       like = CalcLike (Geno, PreGeno, Q, P, Recessive, NULL, NULL); 
     } else {
-			printf ("here 2");
       like = recomblikelihood;
     }
 
     if (rep >= BURNIN + 2) { /* +2 because need 2 observations for variance*/
-			printf ("here 3");
       logprob = EstLogProb (sumlikes, sumsqlikes, rep - BURNIN);
     } else if (COMPUTEPROB) { /*not linkage model*/
       if (rep <= BURNIN + 2) {
-				printf ("here 4");
         like = CalcLike (Geno, PreGeno, Q, P, Recessive, NULL, NULL);
       } else {
-				printf ("here 5");
         logprob = EstLogProb (sumlikes, sumsqlikes, rep - BURNIN);
-				printf ("here 6");
       }
     }
   }
@@ -284,16 +275,22 @@ PrintUpdate (int rep, int *Geno, int *PreGeno, double *Alpha, double *Fst, doubl
   }
 	*/
 
-	double like2=like; /* create new like2 variable in case like is used elsewhere */
+	
+	/* create new like2 variable in case like is used elsewhere */
 	if (rep < (BURNIN+2)) {
-		like2=CalcLike (Geno, PreGeno, Q, P, Recessive, NULL, NULL);
+		like2 = CalcLike (Geno, PreGeno, Q, P, Recessive, NULL, NULL);
 	} else {
-		like2=like;
+		like2 = like;
 	}
 	printf ("  %.0f  ", like2);
 	printf ("  %.0f ", logprob);
   printf ("\n");
+	
+	
 	/* end edits */
+	
+	
+	
 	
   if (rep == BURNIN) {
     printf ("\nBURNIN completed");