+++ NEWS.md
@@ -1,3 +1,9 @@
+# structurer 1.0.1
+
+* Added functionality to check convergence of Structure runs
++ added methods to make traceplots of MCMC runs
++ added methods to calculate Gelman-Rubin statistics
+
 # structurer 1.0.0
 
 * Basic functionality